wandb/run-20260309_213953-z85rfij0/files/requirements.txt
triton==3.6.0
nvidia-cuda-runtime-cu12==12.8.90
sympy==1.14.0
Jinja2==3.1.6
nvidia-cufft-cu12==11.3.3.83
nvidia-cublas-cu12==12.8.4.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-curand-cu12==10.3.9.90
typing_extensions==4.15.0
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusolver-cu12==11.7.3.90
nvidia-nccl-cu12==2.27.5
cuda-pathfinder==1.4.0
nvidia-cuda-cupti-cu12==12.8.90
mpmath==1.3.0
networkx==3.6.1
nvidia-nvtx-cu12==12.8.90
nvidia-cufile-cu12==1.13.1.3
MarkupSafe==3.0.3
fsspec==2026.2.0
torch==2.10.0
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
cuda-bindings==12.9.4
nvidia-cusparselt-cu12==0.7.1
nvidia-cudnn-cu12==9.10.2.21
joblib==1.5.3
tqdm==4.67.3
kiwisolver==1.4.9
filelock==3.25.0
gitdb==4.0.12
tensorboard==2.20.0
captum==0.8.0
Werkzeug==3.1.6
pyparsing==3.3.2
GitPython==3.1.46
annotated-types==0.7.0
contourpy==1.3.3
charset-normalizer==3.4.4
scipy==1.17.1
tensorboard-data-server==0.7.2
six==1.17.0
typing-inspection==0.4.2
pydantic_core==2.41.5
platformdirs==4.9.2
torchvision==0.25.0
certifi==2026.2.25
cycler==0.12.1
requests==2.32.5
matplotlib==3.10.8
pip==26.0.1
torchaudio==2.10.0
wandb==0.25.0
pydantic==2.12.5
pillow==12.1.1
PyYAML==6.0.3
scikit-learn==1.8.0
idna==3.11
urllib3==2.6.3
ImageIO==2.37.2
click==8.3.1
sentry-sdk==2.53.0
absl-py==2.4.0
pandas==3.0.1
setuptools==82.0.0
packaging==26.0
grpcio==1.78.0
seaborn==0.13.2
smmap==5.0.2
fonttools==4.61.1
python-dateutil==2.9.0.post0
Markdown==3.10.2
numpy==1.26.4
wheel==0.46.3
threadpoolctl==3.6.0
protobuf==6.33.5
